fix for error: column permission.createdBy does not exist
sails-permissions initializes some permissions objects with fixture data,
which includes createdBy fields.  It also sets the permissions model
'autoCreatedBy' to false for some reason.  Hence the error.
I just deleted the 'autoCreatedBy: false' config in the permission model.
It starts up with no errors now, but I'm not sure if that was set
to false for a reason.  The tests all still pass.